change logger to be initailised before event manager is used in health check
